--- src/ui/Browser/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Browser";
