+++ src/ui/Project/Project.css
@@ -1,5 +1,5 @@
 
-.Browser .Audiosets .AudiosetReference {
+.Project .Audiosets .AudiosetReference {
   display: flex;
   flex-direction: row;
   align-self: center;
@@ -7,7 +7,7 @@
   margin-bottom: 10px;
   padding-right: 1rem;
 }
-.Browser .Audiosets .AudiosetReference .image {
+.Project .Audiosets .AudiosetReference .image {
   width: 34%;
   height: auto;
   max-width: 256px;
@@ -15,13 +15,13 @@
   flex-shrink: 0  ;
   margin-right: 1rem;
 }
-.Browser .Audiosets .AudiosetReference p {
+.Project .Audiosets .AudiosetReference p {
   max-height: 4.5rem;
   overflow: hidden;
 }
-.Browser .visuals {
+.Project .visuals {
 }
-.Browser .visuals .Markdown {
+.Project .visuals .Markdown {
   padding: 1rem;
   height: 100%;
 }
