+++ src/ui/Project/index.tsx
@@ -1,0 +1,1 @@
+export * from "./Project";
